Interleague Changes

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/xendy/SaveState.java
@@ -7,11 +7,11 @@ public class SaveState implements Serializable {
     public double mX, mY, yaw, mS, turnPower;
 
     // Horizontal Arm
-    public double horizArmPosition;
-    public double vertArmPosition;
+    public int horizArmPosition;
+    public int vertArmPosition;
     public double clawPosition;
     public double bucketPosition;
-    SaveState(double horzPow, double vertPow, double cYaw, double time, double maxSpeed, double turnPow, double harmpos, double varmpos, double bucket, double claw) {
+    SaveState(double horzPow, double vertPow, double cYaw, double time, double maxSpeed, double turnPow, int harmpos, int varmpos, double bucket, double claw) {
         t = time;
         mX = horzPow;
         mY = vertPow;
